requirements.txt
pandas
numpy 
python-dotenv
pymongo
certifi
pymongo[srv]==3.12

## -e . The e . will refer to setup.py